Switch to full with commit cards and disable fast scroll
Re-arrange commit cards to remove the padding around them
Star change is now accessible from the overflow menu
