+++ app/src/main/res/layout/commit_card.xml
@@ -25,8 +25,7 @@
     android:layout_height="wrap_content"
     android:background="?attr/clickableCardBackground"
     android:elevation="@dimen/z_offset_medium"
-    android:orientation="horizontal"
-    tools:showIn="@layout/commit_card_row">
+    android:orientation="horizontal">
 
     <View
         android:id="@+id/commit_card_commit_status"
@@ -39,7 +38,8 @@
     <RelativeLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:paddingBottom="@dimen/standard_padding">
+        android:paddingBottom="@dimen/standard_padding"
+        android:descendantFocusability="blocksDescendants">
 
         <de.hdodenhof.circleimageview.CircleImageView
             android:id="@+id/commit_card_committer_image"
@@ -64,49 +64,64 @@
             tools:text="John Smith" />
 
         <TextView
+            android:id="@+id/commit_card_last_updated"
+            style="@style/CardText"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            tools:text="12:00 AM"
+            android:layout_alignTop="@+id/commit_card_committer_image"
+            android:layout_alignParentRight="true"
+            android:layout_alignParentEnd="true"
+            android:layout_marginRight="@dimen/standard_padding"
+            android:gravity="right"/>
+
+        <TextView
             android:id="@+id/commit_card_title"
             style="?attr/cardTitleStyle"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_toRightOf="@id/commit_card_committer_image"
             android:paddingBottom="@dimen/standard_padding"
             android:paddingRight="@dimen/standard_padding"
             android:textSize="@dimen/text_size_large"
-            tools:text="Tweak commit card layout" />
+            tools:text="Tweak commit card layout"
+            android:layout_alignTop="@+id/commit_card_committer_image"
+            android:layout_toLeftOf="@id/commit_card_last_updated"
+            android:layout_toRightOf="@+id/commit_card_commit_owner"
+            android:layout_toEndOf="@+id/commit_card_commit_owner" />
 
-        <LinearLayout
-            android:layout_width="match_parent"
+        <ImageButton
+            android:layout_width="44dp"
+            android:layout_height="44dp"
+            android:src="?attr/overflowIcon"
+            style="?attr/borderlessButtonBackground"
+            android:contentDescription="@string/menu_overflow_description"
+            android:id="@+id/commit_card_settings"
+            android:layout_gravity="bottom"
+            android:layout_alignBottom="@+id/commit_card_project_name"
+            android:layout_alignParentRight="true"
+            android:layout_alignParentEnd="true"/>
+
+        <TextView
+            android:id="@+id/commit_card_project_name"
+            style="@style/CardText"
+            android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_alignParentBottom="true"
+            android:layout_marginTop="6dp"
+            android:textSize="@dimen/text_size_small"
+            tools:text="AOKP/external_jbirdvegas_mGerrit"
+            android:layout_marginLeft="@dimen/standard_padding"
+            android:layout_toRightOf="@id/commit_card_commit_owner"
             android:layout_below="@id/commit_card_title"
-            android:layout_toEndOf="@id/commit_card_committer_image"
-            android:layout_toRightOf="@id/commit_card_committer_image"
-            android:orientation="horizontal"
-            android:weightSum="1">
+            android:layout_toLeftOf="@id/commit_card_settings"/>
 
-            <TextView
-                android:id="@+id/commit_card_project_name"
-                style="@style/CardText"
-                android:layout_width="0dp"
-                android:layout_height="wrap_content"
-                android:layout_marginTop="6dp"
-                android:layout_weight="1"
-                android:textSize="@dimen/text_size_small"
-                tools:text="AOKP/external_jbirdvegas_mGerrit" />
-
-            <ImageButton
-                android:id="@+id/commit_card_starred"
-                style="?attr/borderlessButtonBackground"
-                android:layout_width="44dp"
-                android:layout_height="44dp"
-                android:layout_gravity="bottom"
-                android:layout_marginBottom="-8dp"
-                android:layout_marginLeft="2dp"
-                android:layout_marginRight="-2dp"
-                android:contentDescription="@string/commit_card_star_change"
-                android:padding="0dp"
-                android:src="?attr/unstarredIcon" />
-        </LinearLayout>
+        <ImageButton
+            android:id="@+id/commit_card_starred"
+            style="?attr/borderlessButtonBackground"
+            android:layout_width="32dp"
+            android:layout_height="32dp"
+            android:layout_gravity="bottom"
+            android:contentDescription="@string/menu_star_change"
+            android:src="?attr/unstarredIcon"/>
     </RelativeLayout>
 
 </com.jbirdvegas.mgerrit.cards.CommitCard>
