--- app/src/main/res/layout/commit_card_row.xml
@@ -1,55 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<!--
-  ~ Copyright (C) 2015 Android Open Kang Project (AOKP)
-  ~  Author: Jon Stanford (JBirdVegas), 2015
-  ~
-  ~  Licensed under the Apache License, Version 2.0 (the "License");
-  ~  you may not use this file except in compliance with the License.
-  ~  You may obtain a copy of the License at
-  ~
-  ~       http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~  Unless required by applicable law or agreed to in writing, software
-  ~  distributed under the License is distributed on an "AS IS" BASIS,
-  ~  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~  See the License for the specific language governing permissions and
-  ~  limitations under the License.
-  -->
-<RelativeLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:orientation="vertical"
-    android:descendantFocusability="blocksDescendants">
-
-    <TextView
-        android:id="@+id/commit_card_last_updated"
-        style="@style/CardText"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginLeft="@dimen/standard_padding"
-        tools:text="12:00 am"
-        android:layout_alignParentLeft="true"
-        android:layout_alignParentStart="true"/>
-
-    <ImageButton
-        android:layout_width="64dp"
-        android:layout_height="wrap_content"
-        android:src="?attr/overflowIcon"
-        style="?attr/borderlessButtonBackground"
-        android:contentDescription="@string/menu_overflow_description"
-        android:id="@+id/commit_card_settings"
-        android:paddingRight="24dp"
-        android:paddingTop="2dp"
-        android:paddingLeft="24dp"
-        android:layout_alignBottom="@+id/commit_card_last_updated"
-        android:layout_alignParentRight="true"
-        android:layout_alignParentEnd="true" />
-
-    <include layout="@layout/commit_card"
-        android:layout_below="@id/commit_card_last_updated"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content" />
-</RelativeLayout>
