+++ app/src/main/res/layout/commit_list.xml
@@ -30,9 +30,9 @@
         android:clipToPadding="false"
         android:divider="@android:color/transparent"
         android:dividerHeight="@dimen/cards_list_divider_height"
-        android:fastScrollEnabled="true"
-        android:scrollbarStyle="outsideInset"
+        android:fastScrollEnabled="false"
+        android:scrollbarStyle="insideOverlay"
+        android:verticalScrollbarPosition="left"
         android:paddingBottom="@dimen/cards_list_padding_vert"
-        android:paddingLeft="@dimen/cards_list_padding_horiz"
         android:drawSelectorOnTop="true"/>
 </android.support.v4.widget.SwipeRefreshLayout>
