+++ app/src/main/res/values/strings.xml
@@ -206,7 +206,6 @@
     <string name="gerrit_signin_username_password_failure">Your username and/or password is incorrect.
         Please try again.</string>
     <string name="gerrit_signin_protected_help">This Gerrit is protected and requires you to sign in before you can view changes. If you do not know your username/password, you can generate one on %s.</string>
-    <string name="commit_card_star_change">Star this change</string>
     <string name="star_change_not_supported" formatted="false">%s does not support starring or unstarring changes in this version. Please ask them to upgrade to version %s or later.</string>
 
     <string name="navigation_drawer_open">Open navigation drawer</string>
@@ -215,6 +214,8 @@
     <string name="menu_view_changes">View Changes</string>
     <string name="menu_starred_changes">Starred Changes</string>
     <string name="menu_post_comment">Post comment</string>
+    <string name="menu_star_change">Star this change</string>
+    <string name="menu_unstar_change">Unstar this change</string>
 
     <string name="review_labels_title">Labels</string>
     <string name="review_advanced_comment_description">Expand to add labels</string>
